split useRoute
