+++ hooks/routing/index.ts
@@ -1,0 +1,18 @@
+// src/hooks/routing/index.ts
+// Export the main hook and all utility types
+
+// Main hook
+export { default as useRoute } from './useRoute';
+
+// Specialized hooks 
+export { useRouteCalculation } from './useRouteCalculation';
+export { useRouteFeatures } from './useRouteFeatures';
+export { useRouteNavigation } from './useRouteNavigation';
+export { useRouteRerouting } from './useRouteRerouting';
+
+// Types
+export * from './utils/types';
+
+// Utils
+export * from './utils/formatters';
+export * from './utils/routeAnalysis';
